Convert shortnames to FQDN during ingestion
