+++ pilot/pkg/config/kube/crd/conversion.go
@@ -36,6 +36,14 @@ func ConvertObject(schema model.ProtoSchema, object IstioObject, domain string) 
 		return nil, err
 	}
 	meta := object.GetObjectMeta()
+
+	// FIXME this is a gross hack to hardcode a service's domain name in kubernetes
+	if domain == "" {
+		domain = "svc.cluster.local"
+	} else {
+		domain = "svc." + domain
+	}
+
 	return &model.Config{
 		ConfigMeta: model.ConfigMeta{
 			Type:            schema.Type,
