+++ pilot/pkg/model/config.go
@@ -556,7 +556,26 @@ func ResolveHostname(meta ConfigMeta, svc *routing.IstioService) string {
 		if svc.Domain != "" {
 			out = out + "." + svc.Domain
 		} else if meta.Domain != "" {
-			out = out + ".svc." + meta.Domain
+			out = out + meta.Domain
+		}
+	}
+
+	return out
+}
+
+// ResolveShortnameToFQDN uses metadata information to resolve a reference
+// to shortname of the service to FQDN
+func ResolveShortnameToFQDN(host string, meta ConfigMeta) string {
+	out := host
+
+	// if FQDN is specified, do not append domain or namespace to hostname
+	if !strings.Contains(host, ".") {
+		if meta.Namespace != "" {
+			out = out + "." + meta.Namespace
+		}
+
+		if meta.Domain != "" {
+			out = out + meta.Domain
 		}
 	}
 
@@ -779,6 +798,30 @@ func (store *istioConfigStore) VirtualServices(gateways []string) []Config {
 		}
 	}
 
+	// Need to parse each rule and convert the shortname to FQDN
+	for _, r := range out {
+		rule := r.Spec.(*networking.VirtualService)
+		// fix top level hosts
+		for i, h := range rule.Hosts {
+			rule.Hosts[i] = ResolveShortnameToFQDN(h, r.ConfigMeta)
+		}
+		// fix http route.destination, route.mirror
+		for _, d := range rule.Http {
+			for _, w := range d.Route {
+				w.Destination.Host = ResolveShortnameToFQDN(w.Destination.Host, r.ConfigMeta)
+			}
+			if d.Mirror != nil {
+				d.Mirror.Host = ResolveShortnameToFQDN(d.Mirror.Host, r.ConfigMeta)
+			}
+		}
+		//fix tcp route.destination
+		for _, d := range rule.Tcp {
+			for _, w := range d.Route {
+				w.Destination.Host = ResolveShortnameToFQDN(w.Destination.Host, r.ConfigMeta)
+			}
+		}
+	}
+
 	return out
 }
 
@@ -852,6 +895,10 @@ func (store *istioConfigStore) DestinationRule(name, domain string) *Config {
 	for _, config := range configs {
 		rule := config.Spec.(*networking.DestinationRule)
 		if ResolveFQDN(rule.Name, domain) == target {
+			return &config
+		}
+		// from shotname destination to FQDN using the rule's namespace
+		if ResolveShortnameToFQDN(rule.Host, config.ConfigMeta) == target {
 			return &config
 		}
 	}
