+++ pilot/pkg/networking/core/v1alpha3/listener.go
@@ -19,7 +19,6 @@ import (
 	"fmt"
 	"os"
 	"sort"
-	"strconv"
 	"strings"
 	"time"
 
@@ -27,7 +26,7 @@ import (
 	"github.com/envoyproxy/go-control-plane/envoy/api/v2/auth"
 	"github.com/envoyproxy/go-control-plane/envoy/api/v2/core"
 	"github.com/envoyproxy/go-control-plane/envoy/api/v2/listener"
-	"github.com/envoyproxy/go-control-plane/envoy/api/v2/route"
+
 	accesslog "github.com/envoyproxy/go-control-plane/envoy/config/filter/accesslog/v2"
 	http_conn "github.com/envoyproxy/go-control-plane/envoy/config/filter/network/http_connection_manager/v2"
 	tcp_proxy "github.com/envoyproxy/go-control-plane/envoy/config/filter/network/tcp_proxy/v2"
@@ -691,190 +690,4 @@ func buildListener(opts buildListenerOpts) *xdsapi.Listener {
 		},
 		DeprecatedV1: deprecatedV1,
 	}
-}
-
-// TODO: find a proper home for these http related functions
-// buildDefaultHTTPRoute builds a default route.
-func buildDefaultHTTPRoute(clusterName string) *route.Route {
-	return &route.Route{
-		Match: route.RouteMatch{PathSpecifier: &route.RouteMatch_Prefix{Prefix: "/"}},
-		Decorator: &route.Decorator{
-			Operation: DefaultOperation,
-		},
-		Action: &route.Route_Route{
-			Route: &route.RouteAction{
-				ClusterSpecifier: &route.RouteAction_Cluster{Cluster: clusterName},
-			},
-		},
-	}
-}
-
-// buildSidecarInboundHTTPRouteConfig builds the route config with a single wildcard virtual host on the inbound path
-// TODO: enable mixer configuration, websockets, trace decorators
-func (configgen *ConfigGeneratorImpl) buildSidecarInboundHTTPRouteConfig(env model.Environment,
-	node model.Proxy, instance *model.ServiceInstance) *xdsapi.RouteConfiguration {
-
-	clusterName := model.BuildSubsetKey(model.TrafficDirectionInbound, "",
-		instance.Service.Hostname, instance.Endpoint.ServicePort)
-	defaultRoute := buildDefaultHTTPRoute(clusterName)
-
-	inboundVHost := route.VirtualHost{
-		Name:    fmt.Sprintf("%s|http|%d", model.TrafficDirectionInbound, instance.Endpoint.ServicePort.Port),
-		Domains: []string{"*"},
-		Routes:  []route.Route{*defaultRoute},
-	}
-
-	// TODO: mixer disabled for now as its configuration is still in old format
-	// set server-side mixer filter config for inbound HTTP routes
-	//if mesh.MixerCheckServer != "" || mesh.MixerReportServer != "" {
-	//	defaultRoute.OpaqueConfig = v1.BuildMixerOpaqueConfig(!mesh.DisablePolicyChecks, false, instance.Service.Hostname)
-	//}
-
-	r := &xdsapi.RouteConfiguration{
-		Name:         clusterName,
-		VirtualHosts: []route.VirtualHost{inboundVHost},
-		ValidateClusters: &google_protobuf.BoolValue{
-			Value: false,
-		},
-	}
-
-	// call plugins
-	for _, p := range configgen.Plugins {
-		p.OnInboundRoute(env, node, instance.Service, instance.Endpoint.ServicePort, r)
-	}
-	return r
-}
-
-func (configgen *ConfigGeneratorImpl) buildSidecarOutboundHTTPRouteConfig(env model.Environment, node model.Proxy,
-	_ []*model.ServiceInstance, services []*model.Service, routeName string) *xdsapi.RouteConfiguration {
-
-	port := 0
-	if routeName != RDSHttpProxy {
-		var err error
-		port, err = strconv.Atoi(routeName)
-		if err != nil {
-			return nil
-		}
-	}
-
-	nameToServiceMap := make(map[string]*model.Service)
-	for _, svc := range services {
-		if port == 0 {
-			nameToServiceMap[svc.Hostname] = svc
-		} else {
-			if svcPort, exists := svc.Ports.GetByPort(port); exists {
-				nameToServiceMap[svc.Hostname] = &model.Service{
-					Hostname: svc.Hostname,
-					Address:  svc.Address,
-					Ports:    []*model.Port{svcPort},
-				}
-			}
-		}
-	}
-
-	// Get list of virtual services bound to the mesh gateway
-	virtualServices := env.VirtualServices([]string{model.IstioMeshGateway})
-	// TODO: Need to trim output based on source label/gateway match
-	guardedHosts := TranslateVirtualHosts(virtualServices, nameToServiceMap)
-	vHostPortMap := make(map[int][]route.VirtualHost)
-
-	for _, guardedHost := range guardedHosts {
-		routes := make([]route.Route, 0)
-		for _, r := range guardedHost.Routes {
-			routes = append(routes, r.Route)
-		}
-
-		virtualHosts := make([]route.VirtualHost, 0)
-
-		for _, host := range guardedHost.Hosts {
-			virtualHosts = append(virtualHosts, route.VirtualHost{
-				Name:    fmt.Sprintf("%s:%d", host, guardedHost.Port),
-				Domains: []string{host},
-				Routes:  routes,
-			})
-		}
-
-		for _, svc := range guardedHost.Services {
-			domains := generateAltVirtualHosts(svc.Hostname, guardedHost.Port)
-			if len(svc.Address) > 0 {
-				// add a vhost match for the IP (if its non CIDR)
-				cidr := util.ConvertAddressToCidr(svc.Address)
-				if cidr.PrefixLen.Value == 32 {
-					domains = append(domains, svc.Address)
-					domains = append(domains, fmt.Sprintf("%s:%d", svc.Address, guardedHost.Port))
-				}
-			}
-			virtualHosts = append(virtualHosts, route.VirtualHost{
-				Name:    fmt.Sprintf("%s:%d", svc.Hostname, guardedHost.Port),
-				Domains: domains,
-				Routes:  routes,
-			})
-		}
-
-		vHostPortMap[guardedHost.Port] = append(vHostPortMap[guardedHost.Port], virtualHosts...)
-	}
-
-	var virtualHosts []route.VirtualHost
-	if routeName == RDSHttpProxy {
-		virtualHosts = mergeAllVirtualHosts(vHostPortMap)
-	} else {
-		virtualHosts = vHostPortMap[port]
-	}
-
-	out := &xdsapi.RouteConfiguration{
-		Name:         fmt.Sprintf("%d", port),
-		VirtualHosts: virtualHosts,
-		ValidateClusters: &google_protobuf.BoolValue{
-			Value: false, // until we have rds
-		},
-	}
-
-	// call plugins
-	for _, p := range configgen.Plugins {
-		p.OnOutboundRoute(env, node, out)
-	}
-
-	return out
-}
-
-// Given a service, and a port, this function generates all possible HTTP Host headers.
-// For example, a service of the form foo.local.campus.net on port 80 could be accessed as
-// http://foo:80 within the .local network, as http://foo.local:80 (by other clients in the campus.net domain),
-// as http://foo.local.campus:80, etc.
-func generateAltVirtualHosts(hostname string, port int) []string {
-	vhosts := []string{hostname, fmt.Sprintf("%s:%d", hostname, port)}
-	for i := len(hostname) - 1; i >= 0; i-- {
-		if hostname[i] == '.' {
-			variant := hostname[:i]
-			variantWithPort := fmt.Sprintf("%s:%d", variant, port)
-			vhosts = append(vhosts, variant)
-			vhosts = append(vhosts, variantWithPort)
-		}
-	}
-	return vhosts
-}
-
-// mergeAllVirtualHosts across all ports. On routes for ports other than port 80,
-// virtual hosts without an explicit port suffix (IP:PORT) should be stripped
-func mergeAllVirtualHosts(vHostPortMap map[int][]route.VirtualHost) []route.VirtualHost {
-	var virtualHosts []route.VirtualHost
-	for p, vhosts := range vHostPortMap {
-		if p == 80 {
-			virtualHosts = append(virtualHosts, vhosts...)
-		} else {
-			for _, vhost := range vhosts {
-				var newDomains []string
-				for _, domain := range vhost.Domains {
-					if strings.Contains(domain, ":") {
-						newDomains = append(newDomains, domain)
-					}
-				}
-				if len(newDomains) > 0 {
-					vhost.Domains = newDomains
-					virtualHosts = append(virtualHosts, vhost)
-				}
-			}
-		}
-	}
-	return virtualHosts
 }
